Modeler UX: Info tab, side-panel indicators, preset no-pause, persist tab state
- Split the Properties panel's Presentation section into a new 'Info' (I) tab
  (InfoPanelContent); Properties keeps Structure / Execution / Indicators.
- Indicators editor now opens in the shared side detail panel (master-detail),
  hosted by Properties via an indicator:<id> selection slot.
- Loading a parameter preset no longer pauses a running simulation; pause only
  when the preset forces a structural worker reinit (grid dims / boundary).
- Preserve Modeler panel/selection state across Modeler<->Simulator tab switches
  via a module-global snapshot (modelerUiState), mirroring graphState's pattern.
- Docs synced: CLAUDE.md, README.md, HelpView.
